+++ src/services/Clients/DeleteClientByIdService.ts
@@ -1,12 +1,17 @@
+import AppError from '@errors/AppError';
 import { getRepository } from 'typeorm';
 
-import Client from '../../models/Client';
+import Client from '@models/Client';
 
 class DeleteClientByIdService {
   public async execute(id: string): Promise<void> {
     const clientsRepository = getRepository(Client);
 
     const client = await clientsRepository.findOne(id);
+
+    if (!client) {
+      throw new AppError('Client does not exists');
+    }
 
     await clientsRepository.remove(client);
   }
